DTO changes

+++ src/main/java/com/socialmedia/socialmediaclone/repository/PostRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
 
     Page<Post> findByUserOrderByDateCreatedDesc(User user, Pageable pageable);
